+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Load ordered sub events.[#48](https://github.com/rokwire/illinois-app/issues/48)
 
 ## [2.3.0] - 2020-09-08
 ### Fixed